// src/core/enums/calculator-button.enum.ts
export enum ITEM {
  CONCAT_PLUS = 'tool-box-item-both',
  LAST = 'tool-box-item-last'
}

export enum ImmediateArithmetic {
  ROOT = '√',
  PERCENT = '%',
  EXACTLY = '=',
  MOD = '+/-'
}

export enum Memory {
  MRC = 'MRC',
  MEMORY_PLUS = 'M+',
  MEMORY_MINUS = 'M-'
}

export enum ArithmeticSign {
  PLUS = '+',
  MINUS = '-',
  INCREASE = 'X',
  DIVISION = '/'
}